# pkg/snowflake: reduce contention in Next calls