@hercules-influx 2018-08-15
During a run of [megacheck](https://github.com/dominikh/go-tools) the following issues were discovered:
```/tmp/610360901/src/github.com/influxdata/influxdb/pkg/snowflake/gen.go:17:2: const sequenceMask is unused (U1000)
```